No expressions bound by infinity, verbosity options
OsiSolverInterface returns two expressions per RowVector one for less
than and one for greater than, this lead to a lot of redundant
expressions bound by infinity on both sides. This occationally made
solvers that don't support integer variables some odd issues.

MIPParser now has a new option '-v or --verbosity' for setting the
verbosity of the solver to be called.