Add error message to get_index_n/p if called without a value list being set.
This may happen if a LeqDomain domain is created, e.g. (x <= 5) == z, and equal() is called without the order encoding of 'x' being enabled.
For now, in the unit tests, it will enable the order encoding where needed, but we should find a better solution for parameterized test-cases in future.